requirements.txt
bcrypt==4.3.0
cffi==1.17.1
click==8.2.0
coolname==2.2.0
crcmod==1.7
cryptography==44.0.3
dearpygui==2.0.0
grpcio==1.71.0
grpcio-tools==1.71.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.5
pandas==2.2.3
paramiko==3.5.1
protobuf==5.29.4
protoletariat==3.3.10
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rich==14.0.0
science-synapse==2.2.0
scipy==1.15.3
six==1.17.0
-e git+ssh://git@github.com/arminvoid/synapse-oephys-driver.git@1c2cb80ba243ae63c54df762c94ce1acd1b2b749#egg=synapse_oephys
tzdata==2025.2
